requirements.txt
cloudinary==1.24.0
dj3-cloudinary-storage==0.0.3
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
google-api-python-client==1.12.8
Pillow==8.0.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
requests==2.25.1
gunicorn==20.0.4
